Build html with js, bake react during build
Produce a static html file that loaded can produce the initial layout
of the game. This can be simply a loading splash, letting it load
before the js can even finish as any browser will load the css and
splash's images first. Or it can contain more behind the splash like
the first screen allowing the browser to load so they might be ready
once the js loads.