Add AnimatedPreview tool
- Lint tools
- Switch dev build to name files [name] instead of [hash]. With just
  [hash] the two output js files were colliding.

AnimatedPreview adds the concepts of armatures and clips to animate
them. Armatures are DOM hierarchies (that can be maintained by React)
with named elements that animation clips animate. As of this changeset
AnimatedPreview holds ClipPlayer and ClipPreviewPlayer (a subclass used
in AnimatedPreview). These classes will shortly go upstream to BoxArt.